refactor: extract mentions tests to mentions component test file
